myblog/src/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
django-taggit==1.3.0
gunicorn==20.0.4
pkg-resources==0.0.0
psycopg2-binary==2.8.5
python-dotenv==0.14.0
pytz==2020.1
slug==2.0
sqlparse==0.3.1
whitenoise==5.1.0
